requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.53
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dateutil==2.8.2
argon2-cffi==23.1.0
